requirements.txt
# --- Flask & Web ---
flask==3.1.1
flask-session==0.5.0
python-dotenv==1.0.1
requests==2.32.3
email-validator==2.1.0.post1

# --- RAG / LLM ---
openai>=1.3.8
langchain==0.1.17
langchain-core==0.1.48
langchain-community==0.0.36
langchain-text-splitters==0.0.1
faiss-cpu==1.11.0

# --- Local Embeddings ---
sentence-transformers==2.5.1

# --- Utilities ---
tiktoken==0.9.0
reportlab==4.4.1
pypdf==3.17.4
tenacity==8.2.3